+++ djangoql/static/djangoql/js/completion.js
@@ -5,17 +5,17 @@
 
   if (typeof define === 'function' && define.amd) {
     // AMD. Register as an anonymous module.
-    define('DjangoQL', ['Lexer'], factory);
+    define('DjangoQL', ['Lexer', 'LRUCache'], factory);
   } else if (typeof exports === 'object') {
     // Node. Does not work with strict CommonJS, but
     // only CommonJS-like environments that support module.exports,
     // like Node.
-    module.exports = factory(require('Lexer'));  // eslint-disable-line
+    module.exports = factory(require('Lexer'), require('LRUCache'));  // eslint-disable-line
   } else {
     // Browser globals (root is window)
-    root.DjangoQL = factory(root.Lexer);  // eslint-disable-line
+    root.DjangoQL = factory(root.Lexer, root.LRUCache);  // eslint-disable-line
   }
-}(this, function (Lexer) {
+}(this, function (Lexer, LRUCache) {
   'use strict';
 
   var reIntValue = '(-?0|-?[1-9][0-9]*)';
@@ -112,9 +112,12 @@
 
   // Main DjangoQL object
   var DjangoQL = function (options) {
+    var cacheSize = 100;
+
     this.options = options;
     this.currentModel = null;
     this.models = {};
+    this.suggestionsAPIUrl = null;
 
     this.token = token;
     this.lexer = lexer;
@@ -153,6 +156,19 @@
     if (options.valuesCaseSensitive) {
       this.valuesCaseSensitive = true;
     }
+    if (options.cacheSize) {
+      if (parseInt(options.cacheSize, 10) !== options.cacheSize
+          || options.cacheSize < 1) {
+        this.logError('cacheSize must be a positive integer');
+      } else {
+        cacheSize = options.cacheSize;
+      }
+    }
+    this.suggestionsCache = new LRUCache(cacheSize);
+    this.debouncedLoadFieldOptions = this.debounce(
+      this.loadFieldOptions.bind(this),
+      300);
+    this.loading = false;
 
     this.enableCompletion = this.enableCompletion.bind(this);
     this.disableCompletion = this.disableCompletion.bind(this);
@@ -220,6 +236,9 @@
         this.completion.className = 'djangoql-completion';
         document.querySelector('body').appendChild(this.completion);
         this.completionUL = document.createElement('ul');
+        this.completionUL.onscroll = this.throttle(
+          this.onCompletionScroll.bind(this),
+          50);
         this.completion.appendChild(this.completionUL);
         if (typeof options.syntaxHelp === 'string') {
           syntaxHelp = document.createElement('p');
@@ -258,37 +277,43 @@
       this.hideCompletion();
     },
 
+    getJson: function (url, settings) {
+      var onLoadError = function () {
+        this.logError('failed to fetch from ' + url);
+      }.bind(this);
+      var request = new XMLHttpRequest();
+      request.open('GET', url, true);
+      request.onload = function () {
+        if (request.status === 200) {
+          if (typeof settings.success === 'function') {
+            settings.success(JSON.parse(request.responseText));
+          }
+        } else {
+          onLoadError();
+        }
+      };
+      request.ontimeout = onLoadError;
+      request.onerror = onLoadError;
+      /* eslint-disable max-len */
+      // Workaround for IE9, see
+      // https://cypressnorth.com/programming/internet-explorer-aborting-ajax-requests-fixed/
+      /* eslint-enable max-len */
+      request.onprogress = function () {};
+      window.setTimeout(request.send.bind(request));
+    },
+
     loadIntrospections: function (introspections) {
-      var onLoadError;
-      var request;
+      var initIntrospections = function (data) {
+        this.currentModel = data.current_model;
+        this.models = data.models;
+        this.suggestionsAPIUrl = data.suggestions_api_url;
+      }.bind(this);
+
       if (typeof introspections === 'string') {
         // treat as URL
-        onLoadError = function () {
-          this.logError('failed to load introspections from ' + introspections);
-        }.bind(this);
-        request = new XMLHttpRequest();
-        request.open('GET', introspections, true);
-        request.onload = function () {
-          var data;
-          if (request.status === 200) {
-            data = JSON.parse(request.responseText);
-            this.currentModel = data.current_model;
-            this.models = data.models;
-          } else {
-            onLoadError();
-          }
-        }.bind(this);
-        request.ontimeout = onLoadError;
-        request.onerror = onLoadError;
-        /* eslint-disable max-len */
-        // Workaround for IE9, see
-        // https://cypressnorth.com/programming/internet-explorer-aborting-ajax-requests-fixed/
-        /* eslint-enable max-len */
-        request.onprogress = function () {};
-        window.setTimeout(request.send.bind(request));
+        this.getJson(introspections, { success: initIntrospections });
       } else if (this.isObject(introspections)) {
-        this.currentModel = introspections.current_model;
-        this.models = introspections.models;
+        initIntrospections(introspections);
       } else {
         this.logError(
             'introspections parameter is expected to be either URL or ' +
@@ -338,6 +363,100 @@
         }
         return result;
       };
+    },
+
+    throttle: function (func, wait, options) {
+      // Borrowed from Underscore.js
+      var timeout;
+      var context;
+      var args;
+      var result;
+      var previous = 0;
+      var later;
+      var throttled;
+
+      if (!options) {
+        options = {};
+      }
+
+      later = function () {
+        previous = options.leading === false ? 0 : new Date().getTime();
+        timeout = null;
+        result = func.apply(context, args);
+        if (!timeout) {
+          context = null;
+          args = null;
+        }
+      };
+
+      throttled = function () {
+        var now = new Date().getTime();
+        var remaining;
+
+        if (!previous && options.leading === false) {
+          previous = now;
+        }
+        remaining = wait - (now - previous);
+        context = this;
+        args = arguments;
+        if (remaining <= 0 || remaining > wait) {
+          if (timeout) {
+            window.clearTimeout(timeout);
+            timeout = null;
+          }
+          previous = now;
+          result = func.apply(context, args);
+          if (!timeout) {
+            context = null;
+            args = null;
+          }
+        } else if (!timeout && options.trailing !== false) {
+          timeout = window.setTimeout(later, remaining);
+        }
+        return result;
+      };
+
+      throttled.cancel = function () {
+        window.clearTimeout(timeout);
+        previous = 0;
+        timeout = null;
+        context = null;
+        args = null;
+      };
+
+      return throttled;
+    },
+
+    setUrlParams: function (url, params) {
+      var parts = url.split('?');
+      var path = parts[0];
+      var queryString = parts.slice(1).join('?');
+      var pairs = queryString.split('&');
+      var pair;
+      var key;
+      var value;
+      var i;
+
+      for (key in params) {
+        if (params.hasOwnProperty(key)) {
+          key = encodeURI(key);
+          value = encodeURI(params[key]);
+          i = pairs.length;
+          while (i--) {
+            pair = pairs[i].split('=');
+            if (pair[0] === key) {
+              pair[1] = value;
+              pairs[i] = pair.join('=');
+              break;
+            }
+          }
+          if (i < 0) {
+            pairs.push(key + '=' + value);
+          }
+        }
+      }
+      queryString = pairs.join('&');
+      return queryString ? [path, queryString].join('?') : path;
     },
 
     logError: function (message) {
@@ -512,6 +631,7 @@
       var li;
       var liLen;
       var suggestionsLen;
+      var loadingElement;
 
       if (!this.completionEnabled) {
         this.hideCompletion();
@@ -521,7 +641,7 @@
       if (dontForceDisplay && this.completion.style.display === 'none') {
         return;
       }
-      if (!this.suggestions.length) {
+      if (!this.suggestions.length && !this.loading) {
         this.hideCompletion();
         return;
       }
@@ -569,6 +689,18 @@
         li[liLen].removeEventListener('mouseout', this.onCompletionMouseOut);
         li[liLen].removeEventListener('mouseover', this.onCompletionMouseOver);
         this.completionUL.removeChild(li[liLen]);
+      }
+
+      loadingElement = this.completionUL.querySelector('.djangoql-loading');
+      if (this.loading) {
+        if (!loadingElement) {
+          loadingElement = document.createElement('li');
+          loadingElement.className = 'djangoql-loading';
+          loadingElement.innerHTML = '&nbsp;';
+          this.completionUL.appendChild(loadingElement);
+        }
+      } else if (loadingElement) {
+        this.completionUL.removeChild(loadingElement);
       }
 
       inputRect = this.textarea.getBoundingClientRect();
@@ -681,7 +813,8 @@
           scope = 'value';
           model = resolvedName.model;
           field = resolvedName.field;
-          if (prefix[0] === '"' && this.models[model][field].type === 'str') {
+          if (prefix[0] === '"' && (this.models[model][field].type === 'str'
+              || this.models[model][field].options)) {
             prefix = prefix.slice(1);
           }
         }
@@ -700,17 +833,163 @@
       return { prefix: prefix, scope: scope, model: model, field: field };
     },
 
+    getCurrentFieldOptions: function () {
+      var input = this.textarea;
+      var context = this.getContext(input.value, input.selectionStart);
+      var model = this.models[context.model];
+      var field = context.field && model[context.field];
+      var fieldOptions = {
+        cacheKey: null,
+        context: context,
+        field: field,
+        model: model,
+        options: null
+      };
+
+      if (context.scope !== 'value' || !field || !field.options) {
+        return null;
+      }
+      if (Array.isArray(field.options)) {
+        fieldOptions.options = field.options;
+      } else if (field.options === true) {
+        // Means get via API
+        if (!this.suggestionsAPIUrl) {
+          return null;
+        }
+        fieldOptions.cacheKey = context.model + '.' + context.field
+          + '|' + context.prefix;
+      }
+      return fieldOptions;
+    },
+
+    loadFieldOptions: function (loadMore) {
+      var fieldOptions = this.getCurrentFieldOptions() || {};
+      var context = fieldOptions.context;
+      var cached;
+      var requestUrl;
+      var requestParams;
+
+      if (!fieldOptions.cacheKey) {
+        // The context has likely changed, user's cursor is in another position
+        return;
+      }
+      requestParams = {
+        field: context.model + '.' + context.field,
+        search: context.prefix
+      };
+
+      cached = this.suggestionsCache.get(fieldOptions.cacheKey) || {};
+      if (loadMore && cached.has_next) {
+        requestParams.page = cached.page ? cached.page + 1 : 1;
+      } else if (cached.page) {
+        // At least the first page is already loaded
+        return;
+      }
+
+      cached.loading = true;
+      this.suggestionsCache.set(fieldOptions.cacheKey, cached);
+      // Render 'loading' element
+      this.populateFieldOptions();
+      this.renderCompletion();
+
+      requestUrl = this.setUrlParams(this.suggestionsAPIUrl, requestParams);
+      this.getJson(requestUrl, {
+        success: function (data) {
+          var cached = this.suggestionsCache.get(fieldOptions.cacheKey) || {};
+          if (data.page - 1 !== (cached.page || 0)) {
+            // either pages were loaded out of order,
+            // or cache is no longer exists
+            return;
+          }
+          data.items = (cached.items || []).concat(data.items);
+          this.suggestionsCache.set(fieldOptions.cacheKey, data);
+          this.populateFieldOptions();
+          this.renderCompletion();
+        }.bind(this)
+      });
+    },
+
+    populateFieldOptions: function (loadMore) {
+      var fieldOptions = this.getCurrentFieldOptions() || {};
+      var options = fieldOptions.options;
+      var prefix = fieldOptions.context && fieldOptions.context.prefix;
+      var input = this.textarea;
+      var cached;
+      var snippetBefore;
+      var snippetAfter;
+      var textBefore;
+      var textAfter;
+
+      this.loading = false;
+      if (options) {
+        // filter them locally
+        if (this.valuesCaseSensitive) {
+          options = options.filter(function (item) {
+            // Case-sensitive
+            return item.indexOf(prefix) >= 0;
+          });
+        } else {
+          options = options.filter(function (item) {
+            // Case-insensitive
+            return item.toLowerCase().indexOf(prefix.toLowerCase()) >= 0;
+          });
+        }
+      } else {
+        this.suggestions = [];
+        if (!fieldOptions.cacheKey) {
+          return;
+        }
+        cached = this.suggestionsCache.get(fieldOptions.cacheKey) || {};
+        options = cached.items || [];
+        this.loading = Boolean(cached.has_next);
+        if (!cached.loading
+            && (!cached.page || (loadMore && cached.has_next))) {
+          this.loading = true;
+          this.debouncedLoadFieldOptions(loadMore);
+        }
+        if (!options.length) {
+          // Should we show 'no results' message?
+          return;
+        }
+      }
+
+      textBefore = input.value.slice(
+        0, input.selectionStart - this.prefix.length);
+      textAfter = input.value.slice(input.selectionStart);
+      if (textBefore && textBefore[textBefore.length - 1] === '"') {
+        snippetBefore = '';
+      } else {
+        snippetBefore = '"';
+      }
+      if (textAfter[0] !== '"') {
+        snippetAfter = '" ';
+      } else {
+        snippetAfter = '';
+      }
+
+      this.highlightCaseSensitive = this.valuesCaseSensitive;
+      this.suggestions = options.map(function (f) {
+        return suggestion(f, snippetBefore, snippetAfter);
+      });
+    },
+
+    onCompletionScroll: function () {
+      var rectHeight = this.completionUL.getBoundingClientRect().height;
+      var scrollBottom = this.completionUL.scrollTop + rectHeight;
+      if (scrollBottom > rectHeight
+          && scrollBottom > (this.completionUL.scrollHeight - rectHeight)) {
+        this.populateFieldOptions(true);
+      }
+    },
+
     generateSuggestions: function () {
       var input = this.textarea;
       var context;
       var model;
       var field;
       var suggestions;
-      var snippetBefore;
       var snippetAfter;
       var searchFilter;
-      var textBefore;
-      var textAfter;
 
       if (!this.completionEnabled) {
         this.prefix = '';
@@ -741,10 +1020,6 @@
       model = this.models[context.model];
       field = context.field && model[context.field];
 
-      textBefore = input.value.slice(
-          0, input.selectionStart - this.prefix.length);
-      textAfter = input.value.slice(input.selectionStart);
-
       switch (context.scope) {
         case 'field':
           this.suggestions = Object.keys(model).map(function (f) {
@@ -760,7 +1035,8 @@
               suggestions.push('~');
               suggestions.push('!~');
               snippetAfter = ' "|"';
-            } else if (field.type === 'date' || field.type === 'datetime') {
+            } else if (field.type === 'date' || field.type === 'datetime'
+                       || field.options) {
               snippetAfter = ' "|"';
             }
             Array.prototype.push.apply(suggestions, ['>', '>=', '<', '<=']);
@@ -769,7 +1045,8 @@
             return suggestion(s, '', snippetAfter);
           });
           if (field && field.type !== 'bool') {
-            if (['str', 'date', 'datetime'].indexOf(field.type) >= 0) {
+            if (['str', 'date', 'datetime'].indexOf(field.type) >= 0
+                || field.options) {
               snippetAfter = ' ("|")';
             } else {
               snippetAfter = ' (|)';
@@ -788,28 +1065,9 @@
           if (!field) {
             // related field
             this.suggestions = [suggestion('None', '', ' ')];
-          } else if (field.type === 'str') {
-            if (textBefore && textBefore[textBefore.length - 1] === '"') {
-              snippetBefore = '';
-            } else {
-              snippetBefore = '"';
-            }
-            if (textAfter[0] !== '"') {
-              snippetAfter = '" ';
-            } else {
-              snippetAfter = '';
-            }
-            if (!this.valuesCaseSensitive) {
-              searchFilter = function (item) {
-                // Case-insensitive
-                return item.text.toLowerCase()
-                        .indexOf(this.prefix.toLowerCase()) >= 0;
-              }.bind(this);
-            }
-            this.highlightCaseSensitive = this.valuesCaseSensitive;
-            this.suggestions = field.options.map(function (f) {
-              return suggestion(f, snippetBefore, snippetAfter);
-            });
+          } else if (field.options) {
+            this.prefix = context.prefix;
+            this.populateFieldOptions();
           } else if (field.type === 'bool') {
             this.suggestions = [
               suggestion('True', '', ' '),